message.js fixed
