+++ js/api.js
@@ -1,32 +1,3 @@
-// const getData = (url, onSuccess, onFail) => {
-//   fetch(url)
-//     .then((response) => {
-//       if (response.ok) {
-//         return response.json();
-//       }
-//       throw new Error();
-//     })
-//     .then((data) => onSuccess(data))
-//     .catch(() => onFail());
-// };
-
-// const sendData = (url, onSuccess, onFail, body) => {
-//   fetch(url, {
-//     method: 'POST',
-//     body
-//   })
-//     .then((response) => {
-//       if (response.ok) {
-//         return response.json();
-//       }
-//       throw new Error();
-//     })
-//     .then(() => onSuccess())
-//     .catch(() => onFail());
-// };
-
-// export { getData, sendData };
-
 const BASE_URL = 'https://28.javascript.pages.academy/kekstagram';
 const Route = {
   GET_DATA: '/data',
