+++ js/main.js
@@ -1,11 +1,9 @@
-// import { randomPhotos } from './script.js';
 import { renderGallery } from './gallery.js';
 import { getData, sendData } from './api.js';
 import { setOnFormSubmit, closePictureUpload } from './form.js';
 import { showError, showSuccess } from './message.js';
 import { showAlert } from './util.js';
 
-// renderGallery(randomPhotos());
 setOnFormSubmit(async (data) => {
   try {
     await sendData(data);
